+++ src/lib/operation-provider.ts
@@ -6,7 +6,7 @@ import { PlugOperation } from "./plug";
 
 export interface OperationProvider {
 	operateGroup(group: Group, operation: GroupOperation, force?: boolean): Promise<boolean>;
-	operateLight(accessory: Accessory, operation: LightOperation): Promise<boolean>;
-	operatePlug(accessory: Accessory, operation: PlugOperation): Promise<boolean>;
-	operateBlind(accessory: Accessory, operation: BlindOperation): Promise<boolean>;
+	operateLight(accessory: Accessory, operation: LightOperation, force?: boolean): Promise<boolean>;
+	operatePlug(accessory: Accessory, operation: PlugOperation, force?: boolean): Promise<boolean>;
+	operateBlind(accessory: Accessory, operation: BlindOperation, force?: boolean): Promise<boolean>;
 }